Fail test for string with max substring 3
