+++ src/test/java/com/lperthel/datastructures/LongestSubstringWithoutRepeatingCharactersTest.java
@@ -1,0 +1,17 @@
+package com.lperthel.datastructures;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+import org.junit.jupiter.api.Test;
+
+class LongestSubstringWithoutRepeatingCharactersTest {
+private Solution solution = new Solution();
+	@Test
+	void test_GivenStringWithLongestSubstringLengthThree_ExpectThree() {
+		String input = "abcabcbb";
+		int expected = 3;
+		int actual = solution.lengthOfLongestSubstring(input);
+		assertEquals(expected,actual);
+	}
+
+}
