+++ src/test/java/com/lperthel/datastructures/TrappingRainWaterTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertEquals;
 import org.junit.jupiter.api.Test;
 
 class TrappingRainWaterTest {
-private Solution  solution = new Solution ();
+private TrappingRainwater  solution = new TrappingRainwater ();
 	@Test
 	void test_GivenVariableHeightTerran_ExpectAreaOfRain() {
 		int[] height = {0,1,0,2,1,0,1,3,2,1,2,1};
